+++ CHANGELOG.md
@@ -3,8 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
-
 ## [Unreleased]
+
+## [0.2.2] - 2022-01-08
 ### Added
 - Add method `lang()` to change language error message.